Modified nextauth configuration
