+++ .env.example
@@ -14,7 +14,7 @@ AUTH_REDIRECT_PROXY_URL="https://authjs.dev/getting-started/deployment#securing-
 AUTH_GITHUB_ID=""
 AUTH_GITHUB_SECRET=""
 
-# https://console.cloud.google.com/apis/dashboard
+# https://console.developers.google.com/apis/credentials
 AUTH_GOOGLE_ID=""
 AUTH_GOOGLE_SECRET=""
