+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,1 +1,3 @@
-export { GET, POST } from "~/server/auth";
+import { handlers } from "~/server/auth";
+
+export const { GET, POST } = handlers;
