+++ src/middleware.ts
@@ -1,15 +1,8 @@
-import NextAuth from "next-auth";
+export { auth as middleware } from "~/server/auth";
 
-import authConfig from "~/auth.config";
-import {
-  apiAuthPrefix,
-  authRoutes,
-  DEFAULT_LOGIN_REDIRECT,
-  publicRoutes,
-} from "~/routes";
-
-const { auth } = NextAuth(authConfig);
-
+/**
+ * import { auth } from "@/auth"
+ 
 export default auth((req) => {
   const { nextUrl } = req;
   const isLoggedIn = !!req.auth;
@@ -30,19 +23,15 @@ export default auth((req) => {
     return;
   }
 
-  console.log("isLoggedIn", isLoggedIn);
-  console.log("isPublicRoute", isPublicRoute);
-  console.log("isAuthRoute", isAuthRoute);
-  console.log("nextUrl", nextUrl);
-  console.log("new URL:", new URL("/auth/login", nextUrl));
-
   if (!isLoggedIn && !isPublicRoute) {
     return Response.redirect(new URL("/auth/login", nextUrl));
   }
 
   return;
-});
+})
 
 export const config = {
   matcher: ["/((?!.+\\.[\\w]+$|_next).*)", "/", "/(api|trpc)(.*)"],
 };
+ * 
+ */
