+++ src/server/auth.ts
@@ -1,25 +1,55 @@
-import { PrismaAdapter } from "@auth/prisma-adapter";
-import NextAuth from "next-auth";
+/* eslint-disable @typescript-eslint/no-unused-vars */
+import { env } from "process";
 
-import authConfig from "~/auth.config";
-import { findUserById } from "~/server/data-access/user";
+import { PrismaAdapter } from "@auth/prisma-adapter";
+import { compareSync } from "bcrypt-edge";
+import NextAuth from "next-auth";
+import Credentials from "next-auth/providers/credentials";
+import GitHub from "next-auth/providers/github";
+import Google from "next-auth/providers/google";
+
+import { LoginSchema } from "~/schemas/auth";
+import { findUserByEmail, findUserById } from "~/server/data-access/user";
 import { db } from "~/server/db";
 import { updateUserEmailVerified } from "~/server/use-cases/user";
 
-export const {
-  handlers: { GET, POST },
-  auth,
-  signIn,
-  signOut,
-} = NextAuth({
+export const { handlers, signIn, signOut, auth } = NextAuth({
+  adapter: PrismaAdapter(db),
+  providers: [
+    Google,
+    GitHub,
+    Credentials({
+      async authorize(credentials) {
+        const validatedFields = LoginSchema.safeParse(credentials);
+        if (validatedFields.success) {
+          const { email, password } = validatedFields.data;
+
+          try {
+            const user = await findUserByEmail(email);
+            if (!user?.password) {
+              return null;
+            }
+
+            const isPasswordCorrect = compareSync(password, user.password);
+
+            if (isPasswordCorrect) {
+              return user;
+            }
+          } catch (error) {
+            console.error("Error authorizing user credentials:", error);
+          }
+        }
+
+        return null;
+      },
+    }),
+  ],
+  session: {
+    strategy: "jwt",
+  },
   pages: {
     signIn: "/auth/login",
     error: "/auth/error",
-  },
-  events: {
-    async linkAccount({ user }) {
-      await updateUserEmailVerified(user.id!);
-    },
   },
   callbacks: {
     async signIn({ user, account }) {
@@ -60,9 +90,10 @@ export const {
       return token;
     },
   },
-  adapter: PrismaAdapter(db),
-  session: {
-    strategy: "jwt",
+  events: {
+    async linkAccount({ user }) {
+      await updateUserEmailVerified(user.id!);
+    },
   },
-  ...authConfig,
+  debug: true,
 });
